refactor: move account files to modules folder
