+++ src/components/layout/ScreenLayout.tsx
@@ -7,17 +7,18 @@ type ScreenLayoutProps = {
   title: string;
   children: ReactNode;
 };
-const ScreenLayout = ({
+export function ScreenLayout({
   id,
   title,
   children,
-}: ScreenLayoutProps): JSX.Element => (
-  <Stack spacing={2} id={id}>
-    <Typography variant="h2" component="h1">
-      {title}
-    </Typography>
-    <Divider />
-    {children}
-  </Stack>
-);
-export default ScreenLayout;
+}: ScreenLayoutProps): JSX.Element {
+  return (
+    <Stack spacing={2} id={id}>
+      <Typography variant="h2" component="h1">
+        {title}
+      </Typography>
+      <Divider />
+      {children}
+    </Stack>
+  );
+}
