+++ src/modules/README.md
@@ -1,0 +1,7 @@
+# Modules
+
+In this folder we host the main modules for the app:
+
+- Landing: Home page of the project
+- Account: User profile and settings interface
+- Other to come.
