+++ src/routes/account.tsx
@@ -3,7 +3,8 @@ import { Container } from '@mui/material';
 import { Outlet, createFileRoute, redirect } from '@tanstack/react-router';
 
 import { LOGIN_PAGE_PATH } from '@/config/paths';
-import PageWrapper from '@/pages/PageWrapper';
+
+import PageWrapper from '~account/PageWrapper';
 
 export const Route = createFileRoute('/account')({
   beforeLoad: ({ context, location }) => {
