+++ src/routes/account/index.tsx
@@ -2,9 +2,9 @@ import { Stack } from '@mui/material';
 
 import { createFileRoute } from '@tanstack/react-router';
 
-import GraaspIcons from '@/components/main/GraaspIcons';
-import MemberCard from '@/components/main/MemberCard';
-import PersonalizationNotificationCard from '@/components/main/PersonalizationNotificationCard';
+import { MemberCard } from '~account/home/MemberCard';
+import { PlatformSelector } from '~account/home/PlatformSelector';
+import { TipCard } from '~account/home/TipCard';
 
 export const Route = createFileRoute('/account/')({
   component: HomeRoute,
@@ -14,8 +14,8 @@ function HomeRoute() {
   return (
     <Stack gap={4} alignItems="center">
       <MemberCard />
-      <PersonalizationNotificationCard />
-      <GraaspIcons />
+      <TipCard />
+      <PlatformSelector />
     </Stack>
   );
 }
